requirements.txt
# Dependencies
numpy
Pillow
torch
rembg[gpu]
opencv-python